Add GraphicsContext. Add simple triengle rendering in Application init, this will be abstracted later

+++ Fragment/src/Fragment/Events/ApplicationEvent.h
@@ -11,8 +11,8 @@ namespace Fragment
 		WindowResizeEvent(unsigned int width, unsigned int height)
 			: m_Width(width), m_Height(height) { }
 
-		inline float GetWidth() const { return m_Width; }
-		inline float GetHeight() const { return m_Height; }
+		inline unsigned int GetWidth() const { return m_Width; }
+		inline unsigned int GetHeight() const { return m_Height; }
 
 		std::string ToString() const override
 		{
